Extract shared resolve_active_repository helper in git_ui
The repo-resolution logic that respects the workspace's active worktree
override was duplicated across branch_picker, git_picker, and project_diff.
Extract it into a single public free function in git_ui.
